Refactor: move user-account into tlx; push tlx modules into TlxServerComponent
- Merge UserAccountResource and UserAccountWithRegistrationResource into a
  single tlx.user.account.UserAccountResource, exposed by TlxServerComponent
  instead of JudgelsServerComponent.
- Merge the corresponding feign clients into tlx.user.UserAccountClient and
  the integration tests into tlx.api.UserAccountApiIntegrationTests.
- Move MailerModule and UserResetPasswordModule from JudgelsServerComponent
  down into TlxServerComponent (re-scoped @Singleton -> @TlxScope).
- Move judgels.problem.ProblemTagResource accessor to JudgelsServerComponent
  (it has no tlx dependency).

AuthModule stays in JudgelsServerComponent: it is also used by the @Singleton
MichaelComponent, so it cannot be re-scoped to @TlxScope.
